Return errors as appropriate
During large imports, there seems to be some hard barrier at the 30m
aggregate mark (which only takes seconds to hit). At this point all RPCs
start returning a deadline exceeded error. Thus, return a 500 to try the
task again. This *should* speed past the existing feeds and allow others
to run. After some number of failures and retries, all should be in and the
UserData field will be able to be updated. This should be improved, though.