For some reason f is not correctly getting set
Either this is a compiler/runtime bug, or I really misunderstand how this
is supposed to work. But, upon using a debug print, f magically starts
getting set correctly. So it is possibly a go routine timing thing, or a
golang bug itself.

As a fix, just use i, which is always set correctly.