Ohh I found you
This looks culprity: forgot to check for other errors. Theory: the Get
failed and cleared everything in &f, but we proceeded and created empty
Feeds. Fix it twice: check for all errors and use f.Url instead of url so
we guarantee failure if we miss setting f.Url again.