Fix unread story bug
The story published date was settable by feeds. However, this date is not
verified, and could have any date or timezone. And if a feed was only
updated rarely, a user's last read date could be now, and a new story added
with published date in the past, and it would never show up. Instead, use
story created date (a field we always set) as the filter to determine if a
story has been updated.

This requires a move from the index p to c. New story queries will be
wrong until a separate script reputs all stories.