Don't worry about stories being newer than feeds
This should fix the bug where feeds weren't being updated when there was
new content. Always just check all stories. Hopefully costs won't go up
too much. If things are in memcache it won't be too bad.

As something to test: just put everything all the time. I think the
datastore might not do anything on a put of an identical item and thus not
charge.