Try harder to find feeds to update
Evidence suggests that we are up to date. The query for just 100 feeds to
update timeouts after 60s usually, but sometimes works. However with so
many feeds this seems odd. It's possible the feed update averager thing
is working so well that we don't need to update most feeds too often.