Don't unescape or tag strip story titles
If text appears there, it is designed to be displayed. If a feed has tags
in it (like the slashdot games feed, which frequently has </em>, due to a
bug in their feed generator, I believe), this is theoretically by design.
Take the GWT issues feed, which commonly has them.

fixes #226