Try another update idea
The cursor thing was interesting but doesn't appear to pan out - I saw
near 300k tasks in the queue. I suspect the stats don't update often enough
and thus the logic is not dependable. Instead, just insert however many we
expect to process (based on the queue speed) and have a longish cronjob to
get a new batch. This was impossible before due to the app engine timeout
bug. Now that it's working (i.e., big queries), we can dump a bunch at
once.

Bump the info to an error because it should now be much more rare that a
feed is double-updated. Will let it run for a day and see if that turns out
to be true.