Server-side user read date bumping
If there are stories (before applying the story-read filter), use the last
one as the new unread date. If there are no stories, move the date to the
latest feed. Don't use time.Now() because feed updates could have happened
that we don't see. Yes, this will result in the latest feed always getting
queried, but it should guarantee that no stories wrongly get marked as
read.

The unread date now means that a user has read everything before that date.
Thus, feeds and stories with a date equal are unread.