Only reschedule push updates when needed
If, after a push update, the next scheduled time is within 6 hours, assume
that push will tell us, so back off until then. The purpose of the 6 hours
is so that if push is broken it'll fallback to us. However, for updates
that would naturally be longer than 6 hours, let them stay for that time.
This should overall decrease the number of updates: some that were being
shortened to 6 hours won't be anymore, and anything < 6 is still 6.