Limit incoming feed size to 2MB
This should prevent huge feeds from crashing instances due to memory use.
For example, the feed at http://www.rackspace.com is huge. Previously, it
would crash the instance, not allowing it to increment the error counter,
so it would attempt to update every 2 minutes. Now it should slow down.