Check for the UTF-8 xml header before decoding
For feeds without a charset in the content type, this allows them to also
identify their charset via the xml header. We could remove the previous
commit in lieu of this, but the Windows 1252 default is still bad, and
would have gotten triggered if, say, the newline at the end of xml.Header
didn't appear in the xml file.